chore: Moved article sidebar and header to new folder
