+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/Header/ArticleHeader.vue
