+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/Header/EditArticleHeader.vue
