+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/Sidebar/Sidebar.vue
