+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/Sidebar/SidebarHeader.vue
