+++ app/javascript/dashboard/routes/dashboard/helpcenter/components/Sidebar/SidebarSearch.vue
